Renamed firstjoiner to more pythonic name.
