+++ tourney/achievements/first_joiner_achievement.py
